refactor PSM writing
