+++ EngineLayer/PeptideSpectralMatch.cs
@@ -5,25 +5,18 @@ using Proteomics.Fragmentation;
 using Proteomics.ProteolyticDigestion;
 using System;
 using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
 using System.Linq;
-using System.Text;
 
 namespace EngineLayer
 {
     public class PeptideSpectralMatch
     {
-        private const double ToleranceForDoubleResolutionF5 = 1e-6;
-        private const double ToleranceForDoubleResolutionF2 = 1e-3;
-
-        private List<(int Notch, PeptideWithSetModifications Pwsm)> _bestMatchingPeptides;
-
         public const double ToleranceForScoreDifferentiation = 1e-9;
+        private List<(int Notch, PeptideWithSetModifications Pwsm)> _BestMatchingPeptides;
 
         public PeptideSpectralMatch(PeptideWithSetModifications peptide, int notch, double score, int scanIndex, IScan scan, DigestionParams digestionParams, List<MatchedFragmentIon> matchedFragmentIons)
         {
-            _bestMatchingPeptides = new List<(int, PeptideWithSetModifications)>();
+            _BestMatchingPeptides = new List<(int, PeptideWithSetModifications)>();
             ScanIndex = scanIndex;
             FullFilePath = scan.FullFilePath;
             ScanNumber = scan.OneBasedScanNumber;
@@ -41,8 +34,7 @@ namespace EngineLayer
             AddOrReplace(peptide, score, notch, true, matchedFragmentIons);
         }
 
-        // these fields will be null if they are ambiguous
-        public ChemicalFormula ModsChemicalFormula { get; private set; }
+        public ChemicalFormula ModsChemicalFormula { get; private set; } // these fields will be null if they are ambiguous
         public string FullSequence { get; private set; }
         public int? Notch { get; private set; }
         public string BaseSequence { get; private set; }
@@ -54,10 +46,8 @@ namespace EngineLayer
         public string ProteinAccession { get; private set; }
         public string Organism { get; private set; }
         public List<MatchedFragmentIon> MatchedFragmentIons { get; protected set; }
-        public int PsmCount { get; set; }
-
-        // these should never be null under normal circumstances
-        public Dictionary<string, int> ModsIdentified { get; private set; }
+        public int PsmCount { get; internal set; }
+        public Dictionary<string, int> ModsIdentified { get; private set; } // these should never be null under normal circumstances
         public List<double> LocalizedScores { get; internal set; }
         public int ScanNumber { get; }
         public int? PrecursorScanNumber { get; }
@@ -69,22 +59,22 @@ namespace EngineLayer
         public double ScanPrecursorMass { get; }
         public string FullFilePath { get; }
         public int ScanIndex { get; }
-        public int NumDifferentMatchingPeptides { get { return _bestMatchingPeptides.Count; } }
+        public int NumDifferentMatchingPeptides { get { return _BestMatchingPeptides.Count; } }
         public FdrInfo FdrInfo { get; private set; }
         public double Score { get; private set; }
         public double DeltaScore { get; private set; }
         public double RunnerUpScore { get; set; }
         public bool IsDecoy { get; private set; }
         public bool IsContaminant { get; private set; }
-        public readonly DigestionParams DigestionParams;
-        public List<double> AllScores { get; set; }
+        public DigestionParams DigestionParams { get; }
+        public List<double> AllScores { get; internal set; }
         public Dictionary<PeptideWithSetModifications, List<MatchedFragmentIon>> PeptidesToMatchingFragments { get; private set; }
 
         public IEnumerable<(int Notch, PeptideWithSetModifications Peptide)> BestMatchingPeptides
         {
             get
             {
-                return _bestMatchingPeptides.OrderBy(p => p.Item2.FullSequence)
+                return _BestMatchingPeptides.OrderBy(p => p.Item2.FullSequence)
                     .ThenBy(p => p.Item2.Protein.Accession)
                     .ThenBy(p => p.Item2.OneBasedStartResidueInProtein);
             }
@@ -110,8 +100,8 @@ namespace EngineLayer
         {
             if (newScore - Score > ToleranceForScoreDifferentiation) //if new score beat the old score, overwrite it
             {
-                _bestMatchingPeptides.Clear();
-                _bestMatchingPeptides.Add((notch, pwsm));
+                _BestMatchingPeptides.Clear();
+                _BestMatchingPeptides.Add((notch, pwsm));
 
                 if (Score - RunnerUpScore > ToleranceForScoreDifferentiation)
                 {
@@ -125,7 +115,7 @@ namespace EngineLayer
             }
             else if (newScore - Score > -ToleranceForScoreDifferentiation && reportAllAmbiguity) //else if the same score and ambiguity is allowed
             {
-                _bestMatchingPeptides.Add((notch, pwsm));
+                _BestMatchingPeptides.Add((notch, pwsm));
 
                 if (!PeptidesToMatchingFragments.ContainsKey(pwsm))
                 {
@@ -151,10 +141,10 @@ namespace EngineLayer
         public static Dictionary<string, string> DataDictionary(PeptideSpectralMatch psm, IReadOnlyDictionary<string, int> ModsToWritePruned)
         {
             Dictionary<string, string> s = new Dictionary<string, string>();
-            AddBasicMatchData(s, psm);
-            AddPeptideSequenceData(s, psm, ModsToWritePruned);
-            AddMatchedIonsData(s, psm);
-            AddMatchScoreData(s, psm);
+            PsmTsvWriter.AddBasicMatchData(s, psm);
+            PsmTsvWriter.AddPeptideSequenceData(s, psm, ModsToWritePruned);
+            PsmTsvWriter.AddMatchedIonsData(s, psm);
+            PsmTsvWriter.AddMatchScoreData(s, psm);
             return s;
         }
 
@@ -187,27 +177,27 @@ namespace EngineLayer
         /// </summary>
         public void ResolveAllAmbiguities()
         {
-            IsDecoy = _bestMatchingPeptides.Any(p => p.Pwsm.Protein.IsDecoy);
-            IsContaminant = _bestMatchingPeptides.Any(p => p.Pwsm.Protein.IsContaminant);
+            IsDecoy = _BestMatchingPeptides.Any(p => p.Pwsm.Protein.IsDecoy);
+            IsContaminant = _BestMatchingPeptides.Any(p => p.Pwsm.Protein.IsContaminant);
 
-            FullSequence = Resolve(_bestMatchingPeptides.Select(b => b.Pwsm.FullSequence)).ResolvedValue;
-            BaseSequence = Resolve(_bestMatchingPeptides.Select(b => b.Pwsm.BaseSequence)).ResolvedValue;
-            PeptideLength = Resolve(_bestMatchingPeptides.Select(b => b.Pwsm.Length)).ResolvedValue;
-            OneBasedStartResidueInProtein = Resolve(_bestMatchingPeptides.Select(b => b.Pwsm.OneBasedStartResidueInProtein)).ResolvedValue;
-            OneBasedEndResidueInProtein = Resolve(_bestMatchingPeptides.Select(b => b.Pwsm.OneBasedEndResidueInProtein)).ResolvedValue;
-            ProteinLength = Resolve(_bestMatchingPeptides.Select(b => b.Pwsm.Protein.Length)).ResolvedValue;
-            PeptideMonisotopicMass = Resolve(_bestMatchingPeptides.Select(b => b.Pwsm.MonoisotopicMass)).ResolvedValue;
-            ProteinAccession = Resolve(_bestMatchingPeptides.Select(b => b.Pwsm.Protein.Accession)).ResolvedValue;
-            Organism = Resolve(_bestMatchingPeptides.Select(b => b.Pwsm.Protein.Organism)).ResolvedValue;
-            ModsIdentified = Resolve(_bestMatchingPeptides.Select(b => b.Pwsm.AllModsOneIsNterminus)).ResolvedValue;
-            ModsChemicalFormula = Resolve(_bestMatchingPeptides.Select(b => b.Pwsm.AllModsOneIsNterminus.Select(c => (c.Value)))).ResolvedValue;
-            Notch = Resolve(_bestMatchingPeptides.Select(b => b.Notch)).ResolvedValue;
+            FullSequence = PsmTsvWriter.Resolve(_BestMatchingPeptides.Select(b => b.Pwsm.FullSequence)).ResolvedValue;
+            BaseSequence = PsmTsvWriter.Resolve(_BestMatchingPeptides.Select(b => b.Pwsm.BaseSequence)).ResolvedValue;
+            PeptideLength = PsmTsvWriter.Resolve(_BestMatchingPeptides.Select(b => b.Pwsm.Length)).ResolvedValue;
+            OneBasedStartResidueInProtein = PsmTsvWriter.Resolve(_BestMatchingPeptides.Select(b => b.Pwsm.OneBasedStartResidueInProtein)).ResolvedValue;
+            OneBasedEndResidueInProtein = PsmTsvWriter.Resolve(_BestMatchingPeptides.Select(b => b.Pwsm.OneBasedEndResidueInProtein)).ResolvedValue;
+            ProteinLength = PsmTsvWriter.Resolve(_BestMatchingPeptides.Select(b => b.Pwsm.Protein.Length)).ResolvedValue;
+            PeptideMonisotopicMass = PsmTsvWriter.Resolve(_BestMatchingPeptides.Select(b => b.Pwsm.MonoisotopicMass)).ResolvedValue;
+            ProteinAccession = PsmTsvWriter.Resolve(_BestMatchingPeptides.Select(b => b.Pwsm.Protein.Accession)).ResolvedValue;
+            Organism = PsmTsvWriter.Resolve(_BestMatchingPeptides.Select(b => b.Pwsm.Protein.Organism)).ResolvedValue;
+            ModsIdentified = PsmTsvWriter.Resolve(_BestMatchingPeptides.Select(b => b.Pwsm.AllModsOneIsNterminus)).ResolvedValue;
+            ModsChemicalFormula = PsmTsvWriter.Resolve(_BestMatchingPeptides.Select(b => b.Pwsm.AllModsOneIsNterminus.Select(c => (c.Value)))).ResolvedValue;
+            Notch = PsmTsvWriter.Resolve(_BestMatchingPeptides.Select(b => b.Notch)).ResolvedValue;
 
             // if the PSM matches a target and a decoy and they are the SAME SEQUENCE, remove the decoy
             if (IsDecoy)
             {
                 bool removedPeptides = false;
-                var hits = _bestMatchingPeptides.GroupBy(p => p.Pwsm.FullSequence);
+                var hits = _BestMatchingPeptides.GroupBy(p => p.Pwsm.FullSequence);
 
                 foreach (var hit in hits)
                 {
@@ -215,7 +205,7 @@ namespace EngineLayer
                     {
                         // at least one peptide with this sequence is a target and at least one is a decoy
                         // remove the decoys with this sequence
-                        _bestMatchingPeptides.RemoveAll(p => p.Pwsm.FullSequence == hit.Key && p.Pwsm.Protein.IsDecoy);
+                        _BestMatchingPeptides.RemoveAll(p => p.Pwsm.FullSequence == hit.Key && p.Pwsm.Protein.IsDecoy);
                         removedPeptides = true;
                     }
                 }
@@ -238,15 +228,15 @@ namespace EngineLayer
         {
             if (IsDecoy)
             {
-                if (_bestMatchingPeptides.Any(p => parsimoniousProteins.Contains(p.Pwsm.Protein) && p.Pwsm.Protein.IsDecoy))
+                if (_BestMatchingPeptides.Any(p => parsimoniousProteins.Contains(p.Pwsm.Protein) && p.Pwsm.Protein.IsDecoy))
                 {
-                    _bestMatchingPeptides.RemoveAll(p => !parsimoniousProteins.Contains(p.Item2.Protein));
+                    _BestMatchingPeptides.RemoveAll(p => !parsimoniousProteins.Contains(p.Item2.Protein));
                 }
                 // else do nothing
             }
             else
             {
-                _bestMatchingPeptides.RemoveAll(p => !parsimoniousProteins.Contains(p.Item2.Protein));
+                _BestMatchingPeptides.RemoveAll(p => !parsimoniousProteins.Contains(p.Item2.Protein));
             }
 
             ResolveAllAmbiguities();
@@ -257,436 +247,8 @@ namespace EngineLayer
         /// </summary>
         public void AddProteinMatch((int, PeptideWithSetModifications) peptideWithNotch)
         {
-            _bestMatchingPeptides.Add(peptideWithNotch);
+            _BestMatchingPeptides.Add(peptideWithNotch);
             ResolveAllAmbiguities();
-        }
-
-        /// <summary>
-        /// Resolve Methods()
-        /// if all 'values' are the same this returns the one value otherwise you get a separated list of all values in their original order.
-        /// for example:
-        /// Notches 1,1,1,1 returns as 1
-        /// Notches 1,0,1,0 returns as 1|0|1|0
-        /// </summary>
-        private static (string ResolvedString, ChemicalFormula ResolvedValue) Resolve(IEnumerable<IEnumerable<Modification>> enumerable)
-        {
-            var list = enumerable.ToList();
-            ChemicalFormula firstChemFormula = new ChemicalFormula();
-            foreach (var firstMods in list[0])
-            {
-                if (firstMods == null || firstMods.ChemicalFormula == null)
-                {
-                    return ("unknown", null);
-                }
-                firstChemFormula.Add(firstMods.ChemicalFormula);
-            }
-
-            bool equals = true;
-            List<ChemicalFormula> formulas = new List<ChemicalFormula>();
-            foreach (var anEnum in list)
-            {
-                ChemicalFormula fhere = new ChemicalFormula();
-                foreach (var mod in anEnum)
-                {
-                    if (mod == null || mod.ChemicalFormula == null)
-                    {
-                        return ("unknown", null);
-                    }
-                    fhere.Add(mod.ChemicalFormula);
-                }
-                if (!firstChemFormula.Equals(fhere))
-                {
-                    equals = false;
-                }
-                formulas.Add(fhere);
-            }
-            if (!equals)
-            {
-                var returnString = GlobalVariables.CheckLengthOfOutput(string.Join("|", formulas.Select(b => b.Formula)));
-                return (returnString, null);
-            }
-            else
-            {
-                return (firstChemFormula.Formula, firstChemFormula);
-            }
-        }
-
-        private static (string ResolvedString, Dictionary<string, int> ResolvedValue) Resolve(IEnumerable<Dictionary<int, Modification>> enumerable)
-        {
-            var list = enumerable.ToList();
-            Dictionary<string, int> firstDict = list[0].Values.OrderBy(b => b.IdWithMotif).GroupBy(b => b.IdWithMotif).ToDictionary(b => b.Key, b => b.Count());
-
-            bool equals = true;
-            foreach (var dict in list)
-            {
-                Dictionary<string, int> okTest = dict.Values.OrderBy(b => b.IdWithMotif).GroupBy(b => b.IdWithMotif).ToDictionary(b => b.Key, b => b.Count());
-                if (!firstDict.SequenceEqual(okTest))
-                {
-                    equals = false;
-                    break;
-                }
-            }
-            if (!equals)
-            {
-                var returnString = string.Join("|", list.Select(b => string.Join(" ", b.Values.Select(c => c.IdWithMotif).OrderBy(c => c))));
-                returnString = GlobalVariables.CheckLengthOfOutput(returnString);
-                return (returnString, null);
-            }
-            else
-            {
-                return (string.Join(" ", list[0].Values.Select(c => c.IdWithMotif).OrderBy(c => c)), firstDict);
-            }
-        }
-
-        private static (string ResolvedString, double? ResolvedValue) ResolveF2(IEnumerable<double> enumerable)
-        {
-            var list = enumerable.ToList();
-            if (list.Max() - list.Min() < ToleranceForDoubleResolutionF2)
-            {
-                return (list.Average().ToString("F2", CultureInfo.InvariantCulture), list.Average());
-            }
-            else
-            {
-                var returnString = GlobalVariables.CheckLengthOfOutput(string.Join("|", list.Select(b => b.ToString("F2", CultureInfo.InvariantCulture))));
-                return (returnString, null);
-            }
-        }
-
-        private static (string ResolvedString, double? ResolvedValue) Resolve(IEnumerable<double> enumerable)
-        {
-            var list = enumerable.ToList();
-            if (list.Max() - list.Min() < ToleranceForDoubleResolutionF5)
-            {
-                return (list.Average().ToString("F5", CultureInfo.InvariantCulture), list.Average());
-            }
-            else
-            {
-                var returnString = GlobalVariables.CheckLengthOfOutput(string.Join("|", list.Select(b => b.ToString("F5", CultureInfo.InvariantCulture))));
-                return (returnString, null);
-            }
-        }
-
-        private static (string ResolvedString, int? ResolvedValue) Resolve(IEnumerable<int> enumerable)
-        {
-            var list = enumerable.ToList();
-            var first = list[0];
-            if (list.All(b => first.Equals(b)))
-            {
-                return (first.ToString(CultureInfo.InvariantCulture), first);
-            }
-            else
-            {
-                var returnString = GlobalVariables.CheckLengthOfOutput(string.Join("|", list.Select(b => b.ToString(CultureInfo.InvariantCulture))));
-                return (returnString, null);
-            }
-        }
-
-        private static (string ResolvedString, string ResolvedValue) Resolve(IEnumerable<string> enumerable)
-        {
-            var list = enumerable.ToList();
-            string first = list.FirstOrDefault(b => b != null);
-            // Only first if list is either all null or all equal to the first
-            if (list.All(b => b == null) || list.All(b => first.Equals(b)))
-            {
-                return (first, first);
-            }
-            else
-            {
-                var returnString = GlobalVariables.CheckLengthOfOutput(string.Join("|", list));
-                return (returnString, null);
-            }
-        }
-
-        private static (string ResolvedString, string ResolvedValue) Resolve(IEnumerable<string> enumerable, string ambiguousIfNull)
-        {
-            var list = enumerable.ToList();
-            string first = list.FirstOrDefault(b => b != null);
-            // Only first if list is either all null or all equal to the first
-            if (list.All(b => b == null) || list.All(b => first.Equals(b)))
-            {
-                return (first, first);
-            }
-            // use only distinct names if all of the base sequences are the same
-            else if (ambiguousIfNull != null)
-            {
-                var returnString = GlobalVariables.CheckLengthOfOutput(string.Join("|", list.Distinct()));
-                return (returnString, null);
-            }
-            else
-            {
-                var returnString = GlobalVariables.CheckLengthOfOutput(string.Join("|", list));
-                return (returnString, null);
-            }
-        }
-
-        private static void AddBasicMatchData(Dictionary<string, string> s, PeptideSpectralMatch psm)
-        {
-            s[PsmTsvReader.FileName] = psm == null ? " " : Path.GetFileNameWithoutExtension(psm.FullFilePath);
-            s[PsmTsvReader.Ms2ScanNumber] = psm == null ? " " : psm.ScanNumber.ToString(CultureInfo.InvariantCulture);
-            s[PsmTsvReader.Ms2ScanRetentionTime] = psm == null ? " " : psm.ScanRetentionTime.ToString("F5", CultureInfo.InvariantCulture);
-            s[PsmTsvReader.NumExperimentalPeaks] = psm == null ? " " : psm.ScanExperimentalPeaks.ToString("F5", CultureInfo.InvariantCulture);
-            s[PsmTsvReader.TotalIonCurrent] = psm == null ? " " : psm.TotalIonCurrent.ToString("F5", CultureInfo.InvariantCulture);
-            s[PsmTsvReader.PrecursorScanNum] = psm == null ? " " : psm.PrecursorScanNumber.HasValue ? psm.PrecursorScanNumber.Value.ToString(CultureInfo.InvariantCulture) : "unknown";
-            s[PsmTsvReader.PrecursorCharge] = psm == null ? " " : psm.ScanPrecursorCharge.ToString("F5", CultureInfo.InvariantCulture);
-            s[PsmTsvReader.PrecursorMz] = psm == null ? " " : psm.ScanPrecursorMonoisotopicPeakMz.ToString("F5", CultureInfo.InvariantCulture);
-            s[PsmTsvReader.PrecursorMass] = psm == null ? " " : psm.ScanPrecursorMass.ToString("F5", CultureInfo.InvariantCulture);
-            s[PsmTsvReader.Score] = psm == null ? " " : psm.Score.ToString("F3", CultureInfo.InvariantCulture);
-            s[PsmTsvReader.DeltaScore] = psm == null ? " " : psm.DeltaScore.ToString("F3", CultureInfo.InvariantCulture);
-            s[PsmTsvReader.Notch] = psm == null ? " " : Resolve(psm.BestMatchingPeptides.Select(p => p.Notch)).ResolvedString;
-            s[PsmTsvReader.DifferentPeakMatches] = psm == null ? " " : psm.NumDifferentMatchingPeptides.ToString("F5", CultureInfo.InvariantCulture);
-        }
-
-        private static void AddPeptideSequenceData(Dictionary<string, string> s, PeptideSpectralMatch psm, IReadOnlyDictionary<string, int> ModsToWritePruned)
-        {
-            bool pepWithModsIsNull = psm == null || psm.BestMatchingPeptides == null || !psm.BestMatchingPeptides.Any();
-
-            List<PeptideWithSetModifications> pepsWithMods = pepWithModsIsNull ? null : psm.BestMatchingPeptides.Select(p => p.Peptide).ToList();
-
-            s[PsmTsvReader.BaseSequence] = pepWithModsIsNull ? " " : Resolve(pepWithModsIsNull ? null : pepsWithMods.Select(b => b.BaseSequence)).ResolvedString;
-            s[PsmTsvReader.FullSequence] = pepWithModsIsNull ? " " : Resolve(pepWithModsIsNull ? null : pepsWithMods.Select(b => b.FullSequence)).ResolvedString;
-            s[PsmTsvReader.EssentialSequence] = pepWithModsIsNull ? " " : Resolve(pepWithModsIsNull ? null : pepsWithMods.Select(b => b.EssentialSequence(ModsToWritePruned))).ResolvedString;
-            s[PsmTsvReader.PsmCount] = pepWithModsIsNull ? " " : psm.PsmCount.ToString();
-            s[PsmTsvReader.Mods] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.AllModsOneIsNterminus)).ResolvedString;
-            s[PsmTsvReader.ModsChemicalFormulas] = pepWithModsIsNull ? " " :
-                Resolve(pepsWithMods.Select(p => p.AllModsOneIsNterminus.Select(v => v.Value))).ResolvedString;
-            s[PsmTsvReader.ModsCombinedChemicalFormula] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.AllModsOneIsNterminus.Select(c => (c.Value as Modification)))).ResolvedString;
-            s[PsmTsvReader.NumVariableMods] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.NumVariableMods)).Item1;
-            s[PsmTsvReader.MissedCleavages] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.MissedCleavages.ToString(CultureInfo.InvariantCulture))).ResolvedString;
-            s[PsmTsvReader.PeptideMonoMass] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.MonoisotopicMass)).ResolvedString;
-            s[PsmTsvReader.MassDiffDa] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => psm.ScanPrecursorMass - b.MonoisotopicMass)).ResolvedString;
-            s[PsmTsvReader.MassDiffPpm] = pepWithModsIsNull ? " " : ResolveF2(pepsWithMods.Select(b => ((psm.ScanPrecursorMass - b.MonoisotopicMass) / b.MonoisotopicMass * 1e6))).ResolvedString;
-            s[PsmTsvReader.ProteinAccession] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.Protein.Accession), psm.FullSequence).ResolvedString;
-            s[PsmTsvReader.ProteinName] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.Protein.FullName), psm.FullSequence).ResolvedString;
-            s[PsmTsvReader.GeneName] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => string.Join(", ", b.Protein.GeneNames.Select(d => $"{d.Item1}:{d.Item2}"))), psm.FullSequence).ResolvedString;
-            s[PsmTsvReader.OrganismName] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.Protein.Organism)).Item1;
-            s[PsmTsvReader.IntersectingSequenceVariations] = pepWithModsIsNull ? " " : 
-                Resolve(pepsWithMods.Select(b => string.Join(", ", b.Protein.AppliedSequenceVariations
-                    .Where(av => IntersectsWithVariation(b, av, false))
-                    .Select(av => SequenceVariantString(b, av))))).ResolvedString;
-            s[PsmTsvReader.IdentifiedSequenceVariations] = pepWithModsIsNull ? " " :
-                Resolve(pepsWithMods.Select(b => string.Join(", ", b.Protein.AppliedSequenceVariations
-                    .Where(av => IntersectsWithVariation(b, av, true))
-                    .Select(av => SequenceVariantString(b, av))))).ResolvedString;
-            s[PsmTsvReader.SpliceSites] = pepWithModsIsNull ? " " :
-                Resolve(pepsWithMods.Select(b => string.Join(", ", b.Protein.SpliceSites
-                    .Where(d => Includes(b, d))
-                    .Select(d => $"{d.OneBasedBeginPosition.ToString()}-{d.OneBasedEndPosition.ToString()}")))).ResolvedString;
-            s[PsmTsvReader.Contaminant] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.Protein.IsContaminant ? "Y" : "N")).Item1;
-            s[PsmTsvReader.Decoy] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.Protein.IsDecoy ? "Y" : "N")).Item1;
-            s[PsmTsvReader.PeptideDesicription] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.PeptideDescription)).Item1;
-            s[PsmTsvReader.StartAndEndResiduesInProtein] = pepWithModsIsNull ? " " : 
-                Resolve(pepsWithMods.Select(b => ($"[{b.OneBasedStartResidueInProtein.ToString(CultureInfo.InvariantCulture)} to {b.OneBasedEndResidueInProtein.ToString(CultureInfo.InvariantCulture)}]")), psm.FullSequence).ResolvedString;
-            s[PsmTsvReader.PreviousAminoAcid] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.PreviousAminoAcid.ToString())).ResolvedString;
-            s[PsmTsvReader.NextAminoAcid] = pepWithModsIsNull ? " " : Resolve(pepsWithMods.Select(b => b.NextAminoAcid.ToString())).ResolvedString;
-
-            string allScores = " ";
-            string theoreticalsSearched = " ";
-            if (!pepWithModsIsNull && psm.FdrInfo != null && psm.FdrInfo.CalculateEValue)
-            {
-                allScores = string.Join(";", psm.AllScores.Select(p => p.ToString("F2", CultureInfo.InvariantCulture)));
-                theoreticalsSearched = psm.AllScores.Count.ToString();
-            }
-
-            s[PsmTsvReader.AllScores] = allScores;
-            s[PsmTsvReader.TheoreticalsSearched] = theoreticalsSearched;
-            s[PsmTsvReader.DecoyContaminantTarget] = pepWithModsIsNull ? " " : psm.IsDecoy ? "D" : psm.IsContaminant ? "C" : "T";
-        }
-
-        /// <summary>
-        /// Determines whether a peptide includes a splice site
-        /// </summary>
-        /// <param name="pep"></param>
-        /// <param name="site"></param>
-        /// <returns></returns>
-        private static bool Includes(PeptideWithSetModifications pep, SpliceSite site)
-        {
-            return pep.OneBasedStartResidueInProtein <= site.OneBasedBeginPosition && pep.OneBasedEndResidueInProtein >= site.OneBasedEndPosition;
-        }
-
-        /// <summary>
-        /// Checks for an intersection between a peptide and applied variant that shows a sequence change.
-        /// </summary>
-        /// <param name="pep"></param>
-        /// <param name="appliedVariation"></param>
-        /// <returns></returns>
-        private static bool IntersectsWithVariation(PeptideWithSetModifications pep, SequenceVariation appliedVariation, bool checkUnique)
-        {
-            // does it intersect? 
-            int intersectOneBasedStart = Math.Max(pep.OneBasedStartResidueInProtein, appliedVariation.OneBasedBeginPosition);
-            int intersectOneBasedEnd = Math.Min(pep.OneBasedEndResidueInProtein, appliedVariation.OneBasedEndPosition);
-            if (intersectOneBasedEnd < intersectOneBasedStart)
-            {
-                return false;
-            }
-            else if (!checkUnique)
-            {
-                return true;
-            }
-            else
-            { 
-                // if the original sequence is too short or long, the intersect of the peptide and variant is unique
-                int intersectSize = intersectOneBasedEnd - intersectOneBasedStart + 1;
-                int variantZeroBasedStart = intersectOneBasedStart - appliedVariation.OneBasedBeginPosition;
-                bool origSeqIsShort = appliedVariation.OriginalSequence.Length - variantZeroBasedStart < intersectSize;
-                bool origSeqIsLong = appliedVariation.OriginalSequence.Length > intersectSize && pep.OneBasedEndResidueInProtein > intersectOneBasedEnd;
-                if (origSeqIsShort || origSeqIsLong)
-                {
-                    return true;
-                }
-
-                // is the variant sequence intersecting the peptide different than the original sequence?
-                string originalAtIntersect = appliedVariation.OriginalSequence.Substring(intersectOneBasedStart - appliedVariation.OneBasedBeginPosition, intersectSize);
-                string variantAtIntersect = appliedVariation.VariantSequence.Substring(intersectOneBasedStart - appliedVariation.OneBasedBeginPosition, intersectSize);
-                return originalAtIntersect != variantAtIntersect;
-            }
-        }
-
-        /// <summary>
-        /// Makes the string representing a detected sequence variation, including any modifications on a variant amino acid
-        /// </summary>
-        /// <param name="p"></param>
-        /// <param name="d"></param>
-        /// <returns></returns>
-        private static string SequenceVariantString(PeptideWithSetModifications p, SequenceVariation applied)
-        {
-            var modsOnVariantOneIsNTerm = p.AllModsOneIsNterminus
-                .Where(kv => kv.Key == 1 && applied.OneBasedBeginPosition == 1 || applied.OneBasedBeginPosition <= kv.Key - 2 + p.OneBasedStartResidueInProtein && kv.Key - 2 + p.OneBasedStartResidueInProtein <= applied.OneBasedEndPosition)
-                .ToDictionary(kv => kv.Key - applied.OneBasedBeginPosition + 1, kv => kv.Value);
-            PeptideWithSetModifications variantWithAnyMods = new PeptideWithSetModifications(p.Protein, p.DigestionParams, applied.OneBasedBeginPosition, applied.OneBasedEndPosition, p.CleavageSpecificityForFdrCategory, p.PeptideDescription, p.MissedCleavages, modsOnVariantOneIsNTerm, p.NumFixedMods);
-            return $"{applied.OriginalSequence}{applied.OneBasedBeginPosition}{variantWithAnyMods.FullSequence}";
-        }
-
-        public static void AddMatchedIonsData(Dictionary<string, string> s, PeptideSpectralMatch psm)
-        {
-            bool nullPsm = (psm == null);
-
-            StringBuilder seriesStringBuilder = new StringBuilder();
-            StringBuilder mzStringBuilder = new StringBuilder();
-            StringBuilder fragmentDaErrorStringBuilder = new StringBuilder();
-            StringBuilder fragmentPpmErrorStringBuilder = new StringBuilder();
-            StringBuilder fragmentIntensityStringBuilder = new StringBuilder();
-            List<StringBuilder> stringBuilders = new List<StringBuilder> { seriesStringBuilder, mzStringBuilder, fragmentDaErrorStringBuilder, fragmentPpmErrorStringBuilder, fragmentIntensityStringBuilder };
-
-            if (!nullPsm)
-            {
-                var matchedIons = psm.MatchedFragmentIons;
-                if (matchedIons == null)
-                {
-                    matchedIons = psm.PeptidesToMatchingFragments.First().Value;
-                }
-
-                // using ", " instead of "," improves human readability
-                const string delimiter = ", ";
-
-                var matchedIonsGroupedByProductType = matchedIons.GroupBy(i => i.NeutralTheoreticalProduct.ProductType).OrderBy(i => i.Key).ToList();
-
-                foreach (var productType in matchedIonsGroupedByProductType)
-                {
-                    var products = productType.OrderBy(p => p.NeutralTheoreticalProduct.TerminusFragment.FragmentNumber)
-                        .ToList();
-
-                    stringBuilders.ForEach(p => p.Append("["));
-
-                    for (int i = 0; i < products.Count; i++)
-                    {
-                        MatchedFragmentIon ion = products[i];
-                        string ionLabel;
-
-                        double massError = ion.Mz.ToMass(ion.Charge) - ion.NeutralTheoreticalProduct.NeutralMass;
-                        double ppmMassError = massError / ion.NeutralTheoreticalProduct.NeutralMass * 1e6;
-
-                        if (ion.NeutralTheoreticalProduct.NeutralLoss == 0)
-                        {
-                            // no neutral loss
-                            ionLabel = ion.NeutralTheoreticalProduct.ProductType + "" + ion.NeutralTheoreticalProduct.TerminusFragment.FragmentNumber + "+" + ion.Charge;
-                        }
-                        else
-                        {
-                            // ion label with neutral loss
-                            ionLabel = "(" + ion.NeutralTheoreticalProduct.ProductType + "" + ion.NeutralTheoreticalProduct.TerminusFragment.FragmentNumber
-                                + "-" + ion.NeutralTheoreticalProduct.NeutralLoss.ToString("F2") + ")" + "+" + ion.Charge;
-                        }
-
-                        // append ion label
-                        seriesStringBuilder.Append(ionLabel);
-
-                        // append experimental m/z
-                        mzStringBuilder.Append(ionLabel + ":" + ion.Mz.ToString("F5"));
-
-                        // append absolute mass error
-                        fragmentDaErrorStringBuilder.Append(ionLabel + ":" + massError.ToString("F5"));
-
-                        // append ppm mass error
-                        fragmentPpmErrorStringBuilder.Append(ionLabel + ":" + ppmMassError.ToString("F2"));
-
-                        // append fragment ion intensity
-                        fragmentIntensityStringBuilder.Append(ionLabel + ":" + ion.Intensity.ToString("F0"));
-
-                        // append delimiter ", "
-                        if (i < products.Count - 1)
-                        {
-                            stringBuilders.ForEach(p => p.Append(delimiter));
-                        }
-                    }
-
-                    // append product type delimiter
-                    stringBuilders.ForEach(p => p.Append("];"));
-                }
-            }
-
-            // save ion series strings to output dictionary
-            s[PsmTsvReader.MatchedIonSeries] = nullPsm ? " " : seriesStringBuilder.ToString().TrimEnd(';');
-            s[PsmTsvReader.MatchedIonMzRatios] = nullPsm ? " " : mzStringBuilder.ToString().TrimEnd(';');
-            s[PsmTsvReader.MatchedIonMassDiffDa] = nullPsm ? " " : fragmentDaErrorStringBuilder.ToString().TrimEnd(';');
-            s[PsmTsvReader.MatchedIonMassDiffPpm] = nullPsm ? " " : fragmentPpmErrorStringBuilder.ToString().TrimEnd(';');
-            s[PsmTsvReader.MatchedIonIntensities] = nullPsm ? " " : fragmentIntensityStringBuilder.ToString().TrimEnd(';');
-
-            // number of matched ions
-            s[PsmTsvReader.MatchedIonCounts] = nullPsm ? " " : psm.MatchedFragmentIons.Count.ToString();
-        }
-
-        private static void AddMatchScoreData(Dictionary<string, string> s, PeptideSpectralMatch peptide)
-        {
-            string localizedScores = " ";
-            string improvementPossible = " ";
-            if (peptide != null && peptide.LocalizedScores != null)
-            {
-                localizedScores = GlobalVariables.CheckLengthOfOutput(("[" + string.Join(",", peptide.LocalizedScores.Select(b => b.ToString("F3", CultureInfo.InvariantCulture))) + "]"));
-                improvementPossible = (peptide.LocalizedScores.Max() - peptide.Score).ToString("F3", CultureInfo.InvariantCulture);
-            }
-            s[PsmTsvReader.LocalizedScores] = localizedScores;
-            s[PsmTsvReader.ImprovementPossible] = improvementPossible;
-
-            string cumulativeTarget = " ";
-            string cumulativeDecoy = " ";
-            string qValue = " ";
-            string cumulativeTargetNotch = " ";
-            string cumulativeDecoyNotch = " ";
-            string qValueNotch = " ";
-            string eValue = " ";
-            string eScore = " ";
-            if (peptide != null && peptide.FdrInfo != null)
-            {
-                cumulativeTarget = peptide.FdrInfo.CumulativeTarget.ToString(CultureInfo.InvariantCulture);
-                cumulativeDecoy = peptide.FdrInfo.CumulativeDecoy.ToString(CultureInfo.InvariantCulture);
-                qValue = peptide.FdrInfo.QValue.ToString("F6", CultureInfo.InvariantCulture);
-                cumulativeTargetNotch = peptide.FdrInfo.CumulativeTargetNotch.ToString(CultureInfo.InvariantCulture);
-                cumulativeDecoyNotch = peptide.FdrInfo.CumulativeDecoyNotch.ToString(CultureInfo.InvariantCulture);
-                qValueNotch = peptide.FdrInfo.QValueNotch.ToString("F6", CultureInfo.InvariantCulture);
-                if (peptide.FdrInfo.CalculateEValue)
-                {
-                    eValue = peptide.FdrInfo.EValue.ToString("F6", CultureInfo.InvariantCulture);
-                    eScore = peptide.FdrInfo.EScore.ToString("F6", CultureInfo.InvariantCulture);
-                }
-            }
-            s[PsmTsvReader.CumulativeTarget] = cumulativeTarget;
-            s[PsmTsvReader.CumulativeDecoy] = cumulativeDecoy;
-            s[PsmTsvReader.QValue] = qValue;
-            s[PsmTsvReader.CumulativeTargetNotch] = cumulativeTargetNotch;
-            s[PsmTsvReader.CumulativeDecoyNotch] = cumulativeDecoyNotch;
-            s[PsmTsvReader.QValueNotch] = qValueNotch;
-            s[PsmTsvReader.EValue] = eValue;
-            s[PsmTsvReader.EScore] = eScore;
         }
     }
 }
